+++ R/CFAxis.R
@@ -52,7 +52,7 @@ CFAxis <- R6::R6Class("CFAxis",
         cat("Group    :", self$group$fullname, "\n")
 
       longname <- self$attribute("long_name")
-      if (nzchar(longname) && longname != self$name)
+      if (!is.na(longname) && longname != self$name)
         cat("Long name:", longname, "\n")
 
       cat("Length   :", self$NCdim$length)
@@ -64,10 +64,10 @@ CFAxis <- R6::R6Class("CFAxis",
     #' @return A 1-row `data.frame` with some details of the axis.
     brief = function() {
       longname <- self$attribute("long_name")
-      if (!nzchar(longname) || longname == self$name) longname <- ""
+      if (is.na(longname) || longname == self$name) longname <- ""
       unlim <- if (self$NCdim$unlim) "U" else ""
       units <- self$attribute("units")
-      if (!nzchar(units)) units <- ""
+      if (is.na(units)) units <- ""
       if (units == "1") units <- ""
 
       data.frame(id = self$dimid, axis = self$orientation, group = self$group$fullname,
@@ -167,6 +167,30 @@ CFAxis <- R6::R6Class("CFAxis",
       if (index > 0L && index <= length(self$lbls))
         self$lbls[[index]]$values
       else NULL
+    },
+
+    #' @description Write the axis to a netCDF file, including its attributes.
+    #' @param nc The handle of the netCDF file opened for writing or a group in
+    #'   the netCDF file. If `NULL`, write to the file or group where the axis
+    #'   was read from (the file must have been opened for writing). If not
+    #'   `NULL`, the handle to a netCDF file or a group therein.
+    #' @return Self, invisibly.
+    write = function(nc = NULL) {
+      h <- if (inherits(nc, "NetCDF")) nc else self$group$handle
+
+      if (inherits(self, "CFAxisScalar")) {
+        RNetCDF::var.def.nc(h, self$name, self$NCvar$vtype, NA)
+      } else {
+        self$NCdim$write(h)
+        RNetCDF::var.def.nc(h, self$name, self$NCvar$vtype, self$name)
+      }
+      self$write_attributes(h, self$name)
+      RNetCDF::var.put.nc(h, self$name, private$get_values())
+
+      if (!is.null(self$bounds))
+        self$bounds$write(h, self$name)
+
+      invisible(self)
     }
   ),
 
@@ -200,6 +224,13 @@ CFAxis <- R6::R6Class("CFAxis",
         if (inherits(value, "CFLabel") && value$length == self$length)
           self$lbls <- append(self$lbls, value)
       }
+    },
+
+    #' @field unlimited (read-only) Logical to indicate if the axis has an
+    #'   unlimited dimension.
+    unlimited = function(value) {
+      if (missing(value))
+        self$NCdim$unlim
     }
   )
 )