+++ R/CFAxisNumeric.R
@@ -9,6 +9,10 @@
 CFAxisNumeric <- R6::R6Class("CFAxisNumeric",
   inherit = CFAxis,
   private = list(
+    get_values = function() {
+      self$values
+    },
+
     dimvalues_short = function() {
       lbls <- self$labels
       nv <- length(self$values)
@@ -50,7 +54,7 @@ CFAxisNumeric <- R6::R6Class("CFAxisNumeric",
       super$print()
 
       units <- self$attribute("units")
-      if (!nzchar(units)) units <- ""
+      if (is.na(units)) units <- ""
       if (units == "1") units <- ""
 
       lbls <- self$labels