+++ R/makeCFObjects.R
@@ -28,19 +28,24 @@ makeMemoryGroup <- function(id, name, fullname, title, history) {
 #' @param length Length of the dimension of the axis.
 #' @param values The dimension values.
 #' @param bounds The bounds of the dimension values, or `NULL` if not available.
-#' @param units The name of the axis units.
+#' @param units A character string with the axis units.
 #'
 #' @return A `CFAxisLongitude` instance.
 #' @export
 makeLongitudeAxis <- function(id, name, group, length, values, bounds = NULL, units = "") {
   var <- NCVariable$new(id, name, group, "NC_DOUBLE", 1L, NULL)
-  var$attributes <- data.frame(id = 0L, name = "units", type = "NC_CHAR", length = nchar(units), value = units)
   dim <- NCDimension$new(-1L, name, length, FALSE)
   axis <- CFAxisLongitude$new(group, var, dim, values)
+  if (nzchar(units))
+    axis$set_attribute("units", "NC_CHAR", units)
   if (!is.null(bounds)) {
-    var <- NCVariable$new(-1L, paste0(name, "_bnds"), group, "NC_DOUBLE", 2L, NULL)
-    axis$bounds <- CFBounds$new(var, bounds)
+    nm <- paste0(name, "_bnds")
+    var <- NCVariable$new(-1L, nm, group, "NC_DOUBLE", 2L, NULL)
+    dim <- NCDimension$new(-1L, "nv", 2L, FALSE)
+    axis$bounds <- CFBounds$new(var, dim, bounds)
+    axis$set_attribute("bounds", "NC_CHAR", nm)
   }
+  axis$set_attribute("axis", "NC_CHAR", "X")
   axis
 }
 
@@ -55,19 +60,24 @@ makeLongitudeAxis <- function(id, name, group, length, values, bounds = NULL, un
 #' @param length Length of the dimension of the axis.
 #' @param values The dimension values.
 #' @param bounds The bounds of the dimension values, or `NULL` if not available.
-#' @param units The name of the axis units.
+#' @param units A character string with the axis units.
 #'
 #' @return A `CFAxisLatitude` instance.
 #' @export
 makeLatitudeAxis <- function(id, name, group, length, values, bounds, units) {
   var <- NCVariable$new(id, name, group, "NC_DOUBLE", 1L, NULL)
-  var$attributes <- data.frame(id = 0, name = "units", type = "NC_CHAR", length = nchar(units), value = units)
   dim <- NCDimension$new(-1L, name, length, FALSE)
   axis <- CFAxisLatitude$new(group, var, dim, values)
+  if (nzchar(units))
+    axis$set_attribute("units", "NC_CHAR", units)
   if (!is.null(bounds)) {
-    var <- NCVariable$new(-1L, paste0(name, "_bnds"), group, "NC_DOUBLE", 2L, NULL)
-    axis$bounds <- CFBounds$new(var, bounds)
+    nm <- paste0(name, "_bnds")
+    var <- NCVariable$new(-1L, nm, group, "NC_DOUBLE", 2L, NULL)
+    dim <- NCDimension$new(-1L, "nv", 2L, FALSE)
+    axis$bounds <- CFBounds$new(var, dim, bounds)
+    axis$set_attribute("bounds", "NC_CHAR", nm)
   }
+  axis$set_attribute("axis", "NC_CHAR", "Y")
   axis
 }
 
@@ -86,9 +96,20 @@ makeLatitudeAxis <- function(id, name, group, length, values, bounds, units) {
 makeTimeAxis <- function(id, name, group, values) {
   var <- NCVariable$new(id, name, group, "NC_DOUBLE", 1L, NULL)
   if (length(values) == 1L)
-    CFAxisScalar$new(group, var, "T", values)
+    axis <- CFAxisScalar$new(group, var, "T", values)
   else {
     dim <- NCDimension$new(-1L, name, length(values), FALSE)
-    CFAxisTime$new(group, var, dim, values)
+    axis <- CFAxisTime$new(group, var, dim, values)
   }
+  axis$set_attribute("units", "NC_CHAR", values$cal$definition)
+  axis$set_attribute("calendar", "NC_CHAR", values$cal$name)
+  axis$set_attribute("axis", "NC_CHAR", "T")
+  if (!isFALSE(values$bounds)) {
+    nm <- paste0(name, "_bnds")
+    var <- NCVariable$new(-1L, nm, group, "NC_DOUBLE", 2L, NULL)
+    dim <- NCDimension$new(-1L, "nv", 2L, FALSE)
+    axis$bounds <- CFBounds$new(var, dim, values$get_bounds())
+    axis$set_attribute("bounds", "NC_CHAR", nm)
+  }
+  axis
 }